[IMP] web interface for extension
