+++ main.py
@@ -1,67 +1,25 @@
-from PIL import Image, ImageDraw, ImageFont
-import logging
+from pillow_extension import text_highlighting_colors
+import gradio as gr
+from gradio.components import Textbox, Image, UploadButton, ColorPicker
 
-class ImageProcessingError(Exception):
-    """Custom exception for image processing errors."""
-    pass
 
-def text_highlighting_colors(
-    image_src,
-    font_src,
-    size_font,
-    position_text,
-    text_origin,
-    text_align,
-    text_color,
-    highlighting_color,
-):
-    """Highlight text in image.
 
-    Args:
-        image_src (str): Path to image.
-        font_src (str): Path to font.
-        size_font (int): Size of font.
-        position_text (tuple): Position of text.
-        text_origin (str): Text to highlight.
-        text_align (str): Text alignment.
-        text_color (tuple): Color of text.
-        highlighting_color (tuple): Color of highlighting.
+app = gr.Interface(
+    fn=text_highlighting_colors,
+    title="Hello",
+    inputs=[
+        Textbox(label="Size"),
+        Image(label="Image", format="jpg", sources=["upload", "webcam", "clipboard"]),
+        UploadButton(label="Font"),
+        Textbox(label="Position"),
+        Textbox(label="Text"),
+        Textbox(label="Align"),
+        ColorPicker(label="Text Color"),
+        ColorPicker(label="Highlighting Color"),
+    ],
+    outputs=["image"],
+)
 
-    """
-    try:
-        image = Image.open(image_src)
-    except FileNotFoundError:
-        logging.error(f"Image file '{image_src}' not found.")
-        raise ImageProcessingError(f"Failed to load image: {image_src}")
-    
-    try:
-        font = ImageFont.truetype(font_src, size_font)
-    except FileNotFoundError:
-        logging.error(f"Font file '{font_src}' not found.")
-        raise ImageProcessingError(f"Failed to load font: {font_src}")
-         
-    draw = ImageDraw.Draw(image)
-    y = 0 
-    bbox = draw.textbbox(position_text, text_origin, font=font,align=text_align)
-    xx,yy,x1 = bbox[0:3]
-    text2 = text_origin.split('\n')
-    for item in text2:
-        font_width, _ = font.getsize(item)
-
-        if text_align == 'right':
-            left, top, right, bottom = draw.textbbox((x1-font_width,yy+y), item, font=font)
-            draw.rectangle((left-5, top-5, right+5, bottom+5), fill=highlighting_color)
-            draw.text((x1-font_width,yy+y), item, font=font, fill=text_color)
-        elif text_align == 'center':
-            left, top, right, bottom = draw.textbbox((xx/2-font_width/2+x1/2,yy+y), item, font=font)
-            draw.rectangle((left-5, top-5, right+5, bottom+5), fill=highlighting_color)
-            draw.text((xx/2-font_width/2+x1/2,yy+y), item, font=font, fill=text_color)
-        else:
-            left, top, right, bottom = draw.textbbox((xx,yy+y), item, font=font)
-            draw.rectangle((left-5, top-5, right+5, bottom+5), fill=highlighting_color)
-            draw.text((xx,yy+y), item, font=font, fill=text_color)
-        y += size_font
-    image.save("image_highlighting_colors.png")
 
 
 if __name__ == "__main__":
@@ -73,6 +31,6 @@ if __name__ == "__main__":
     align = "center"
     text_color = "#FF1827"
     highlighting_color = "#18D8FF"
-    text_highlighting_colors(
-        image, font, size, position, text, align, text_color, highlighting_color
-    )
+    
+
+    app.launch()
